chore (model): Refactor validation logic for sort_by and order in fetchArticles function
Refactor fetchArticles error logic into a separate utility function "validateSortAndOrder".
Improve readability by organising validation logic.
