+++ db/seeds/utils.js
@@ -20,3 +20,29 @@ exports.formatComments = (comments, idLookup) => {
     };
   });
 };
+
+exports.validateSortAndOrder = (sort_by, order) => {
+  const validSortBy = ["created_at", "author", "title", "topic", "votes"];
+  const validOrder = ["asc", "desc"];
+
+  if (!validSortBy.includes(sort_by) && !validOrder.includes(order)) {
+    const err = new Error();
+    err.status = 400;
+    err.msg = "Invalid sort_by and order values";
+    throw err;
+  }
+
+  if (!validSortBy.includes(sort_by)) {
+    const err = new Error();
+    err.status = 400;
+    err.msg = `Invalid sort_by value: ${sort_by}`;
+    throw err;
+  }
+
+  if (!validOrder.includes(order)) {
+    const err = new Error();
+    err.status = 400;
+    err.msg = `Invalid order value: ${order}`;
+    throw err;
+  }
+};
